requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic-settings




PyMuPDF
python-docx



langchain
langchain-core
langchain-community
langchain-openai